requirements.txt
absl-py==2.0.0
accelerate==0.27.2
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bertopic==0.16.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.59
botocore==1.34.59
bpemb==0.3.4
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.1.0
chromedriver-binary==74.0.3729.6.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.3
confection==0.1.4
conllu==4.5.3
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.8
Cython==0.29.37
daphne==4.0.0
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==4.2.1
docopt==0.6.2
editdistance==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
eng-to-ipa==0.0.2
epitran==1.24
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.104.1
fastjsonschema==2.18.1
ffmpy==0.3.1
filelock==3.12.4
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
fsspec==2023.9.2
ftfy==6.1.3
funcy==2.0
gast==0.5.4
gdown==5.1.0
gensim==4.3.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.2.0
gradio_client==0.7.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
hdbscan==0.8.33
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
hydra-core==1.3.2
hyperlink==21.0.0
hypothesis==6.88.4
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
incremental==22.10.0
ipykernel==6.23.0
ipytablewidgets==0.3.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
Janome==0.5.0
jedi==0.18.2
jellyfish==1.0.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaldiio==2.18.0
kenlm==0.2.0
keras==2.14.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.0
lxml==5.1.0
lz4==4.3.2
marisa-trie==1.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
monotonic-align==1.0.0
more-itertools==10.2.0
moviepy==1.0.3
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.0.6
munkres==1.1.4
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
noisereduce==3.0.0
notebook==7.0.5
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.10.0
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.10
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.0
panphon==0.20.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.12.1
platformdirs==4.2.2
plotly==5.22.0
pooch==1.7.0
pptree==3.1
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==23.12.11
pycparser==2.21
pyctcdecode==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pydicom==2.4.4
pydub==0.25.1
Pygments==2.15.1
pygtrie==2.5.0
pyLDAvis==3.4.1
pynndescent==0.5.12
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-docx==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch_revgrad==0.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.12
pyworld==0.3.4
PyYAML==6.0.1
pyzmq==25.0.2
qtconsole==5.4.4
QtPy==2.4.0
rapidfuzz==3.6.2
recordlinkage==0.16
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.0
schedule==1.2.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
segtok==1.5.11
selenium==4.17.2
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.2
sentence-transformers==2.7.0
sentencepiece==0.1.99
service-identity==21.1.0
shellingham==1.5.4
silpa-common==0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
sumy==0.11.0
svgling==0.4.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.18.0.post0
thefuzz==0.22.1
thinc==8.2.3
threadpoolctl==3.1.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
transformer-smaller-training-vocab==0.3.3
transformers==4.35.0
trio==0.24.0
trio-websocket==0.11.1
tweepy==4.14.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.6
underthesea==6.8.0
underthesea_core==1.0.4
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==2.2.1
utils==1.0.2
uvicorn==0.24.0.post1
vaderSentiment==3.3.2
vboxapi==1.0
vega==4.0.0
vinorm==2.0.7
viphoneme==3.0.0
virtualenv==20.26.2
vncorenlp==1.0.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
Wikipedia-API==0.6.0
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
xgboost==1.7.5
xlrd==2.0.1
zope.interface==6.0
